chatbot/requirement.txt
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1

langchain==0.2.1
langchain-core==0.2.3
langchain-community==0.2.1
langchain-groq==0.1.4

pydantic==2.7.1